+++ src/themes/palette.ts
@@ -7,7 +7,14 @@ export type PaletteColor = {
   dark: string;
   main: string;
   light: string;
+  softlight?: string;
   highlight: string;
+  ultraLight?: string;
+};
+
+// successカラー用に拡張した型
+export type SuccessPaletteColor = PaletteColor & {
+  medium: string; // green[200]用
 };
 
 export type PaletteText = {
@@ -54,7 +61,7 @@ export type Palette = {
 };
 
 export const palette: Palette = {
-  white: colors.basic[50],
+  white: colors.basic[0],
   black: colors.basic[900],
   primary: {
     deepDark: colors.blue[700],
@@ -62,6 +69,7 @@ export const palette: Palette = {
     main: colors.blue[500],
     light: colors.blue[200],
     highlight: colors.blue[100],
+    ultraLight: colors.blue[50],
   },
   primaryPale: {
     deepDark: colors.blue[600],
@@ -69,6 +77,7 @@ export const palette: Palette = {
     main: colors.blue[50],
     light: colors.blue[200],
     highlight: colors.blue[100],
+    ultraLight: colors.blue[50],
   },
   basicDark: {
     deepDark: colors.basic[400],
@@ -76,13 +85,16 @@ export const palette: Palette = {
     main: colors.basic[100],
     light: colors.basic[300],
     highlight: colors.basic[100],
+    ultraLight: colors.basic[50],
   },
   success: {
     deepDark: colors.green[700],
     dark: colors.green[600],
     main: colors.green[500],
     light: colors.green[300],
+    softlight: colors.green[200],
     highlight: colors.green[100],
+    ultraLight: colors.green[50],
   },
   warning: {
     deepDark: colors.yellow[900],
@@ -90,6 +102,7 @@ export const palette: Palette = {
     main: colors.yellow[500],
     light: colors.yellow[400],
     highlight: colors.yellow[100],
+    ultraLight: colors.yellow[50],
   },
   danger: {
     deepDark: colors.red[700],
@@ -97,6 +110,7 @@ export const palette: Palette = {
     main: colors.red[500],
     light: colors.red[300],
     highlight: colors.red[100],
+    ultraLight: colors.red[50],
   },
   gray: {
     deepDark: colors.basic[600],
@@ -104,16 +118,17 @@ export const palette: Palette = {
     main: colors.basic[300],
     light: colors.basic[200],
     highlight: colors.basic[100],
+    ultraLight: colors.basic[50],
   },
   text: {
     primary: colors.blue[500],
     secondary: colors.basic[700],
     disabled: colors.basic[400],
     hint: colors.basic[400],
-    white: "#FFFFFF",
+    white: colors.basic[0],
   },
   background: {
-    default: "#FFFFFF",
+    default: colors.basic[0],
     dark: colors.blue[40],
     active: colors.blue[100],
     hint: colors.blue[50],